refactor(reducer): extract RepoGit reduce-block
Move the repo-registry (scan/add/remove/rename), worktree-operation,
auto-detected repo-association (auto-link/auto-unlink/branch+remote-URL
resolution), and inspector + git-status (toggle/refresh/status-timer +
HEAD-watcher lifecycle) cases out of AppReducer.body into a new
repoGitReducer block in Nex/AppReducer+RepoGit.swift, plus the
scheduleAutoLink / scheduleAutoUnlink scheduling helpers and the
cancellation-ID enums + debounce constants they own. domain(of:) gains
a .repoGit arm and stays exhaustive with no default.

The two .workspaces(.element(...RepoAssociation)) interceptions and the
scheduleAutoLink/scheduleAutoUnlink call sites stay in core (they are
.workspaces-family actions); repoRegistry / gitStatuses /
isInspectorVisible stay on AppReducer.State. The graft-stop merge
sequencing in removeRepo / removeWorktreeAssociation /
autoUnlinkUnusedRepos is preserved verbatim. AppReducer.swift 4546 -> 4063.
